neu v3.9.0

+++ controller/admin/appTipsController.js
@@ -1,103 +1,24 @@
 const Tip = require('../../models/applianceTips')
+const Category = require('../../models/category')
 const Joi = require('joi')
 const RecentBlogDTO = require('../../dto/blog/recentBlog')
-const AWS3 = require('@aws-sdk/client-s3')
-const S3Client = require('../../services/S3')
-const { AWS_S3_USER_ACCESS_KEY,AWS_S3_USER_SECRET_ACCESS_KEY,AWS_S3_REGION,AWS_S3_BUCKET_NAME } = require('../../config/index')
 const AWSService = require('../../services/S3Upload')
 
 
 const tipsController = {
     async createTip(req, res, next) {
-        const tipSchema = Joi.object({
-            title: Joi.string().required(),
-            slug: Joi.string().required(),
-            thumbnail: Joi.allow(null).empty(''),
-            category: Joi.string().required(),
-            content: Joi.string().required(),
-          });
-          const { error } = tipSchema.validate(req.body);
-          
-          // 2. if error in validation -> return error via middleware
-          if (error) {
-            return next(error)
-          }
-          
-          // 3. if email or username is already registered -> return an error
-          const {title,slug,thumbnail,category,content } = req.body;
-          
-          const titleInUse = await Tip.exists({ title });        
-          if (titleInUse) {
-            const error = {
-              status: 409, message:'Appliance Tip Already Exist!'
-            }
-            return next(error)
-          }
-          const {response,updateImg} = await AWSService.uploadFile({name:req.files.thumbnail.name,data:req.files.thumbnail.data},'appliance-tips/')
-             
-             if(response.$metadata.httpStatusCode === 200){
-               try{
-                const TipToCreate = new Tip({title,slug,thumbnail:updateImg,category,content});
-                const tip = await TipToCreate.save();
-                return res.status(200).json({status: 200, msg:'Appliance Tip Created Successuly!'});
-               }catch(err){
-                 const error = {status:500,msg:"Internal Server Error!"}
-                 return next(error)
-               }
-             }else{
-              const error = {
-                status: 500,
-                message: "AWS S3 Internal Server Error!"
-              }
-              return next(error)
-             }
-    },
-    async duplicateTip(req, res, next) {
-      const tipSchema = Joi.object({
-          slug: Joi.string().required(),
-        });
-        const { error } = tipSchema.validate(req.body);
-        
-        // 2. if error in validation -> return error via middleware
-        if (error) {
-          return next(error)
-        }
-        
-        // 3. if email or username is already registered -> return an error
-        const {slug} = req.body;
-        
-        const tip = await Tip.findOne({ slug });  
-        if (tip) {
-          const {resp,updateImg} = await AWSService.duplicateFile(tip.thumbnail,'appliance-tips/')
-          if(resp.$metadata.httpStatusCode === 200){
-            const title = tip.title + '(Duplicate)';
-            const slug = title.toLocaleLowerCase().replace(/\s/g,'-');
-            try{
-              const TipToCreate = new Tip({title:title,slug:slug,thumbnail:updateImg,category:tip.category,content:tip.content});
-              await TipToCreate.save();
-              return res.status(200).json({status: 200, msg:'Appliance Tip Duplicate Created!'});
-            }catch(err){
-              const error = {status:500,message:"Internal Server Error!"}
-              return next(error)
-            }
-          }else{
-            return res.status(500).json({status: 500, message:'AWS Cloud Internal Server Error!'});
-          }
-        }else{
-          return res.status(404).json({status: 404, message:'Appliance Tip Does Not Exist!'});
-        }  
-  },
-    async updateTip(req, res, next) {
-      const tipSchema = Joi.object({
-          id: Joi.string().required(),
+      const blogSchema = Joi.object({
           title: Joi.string().required(),
           slug: Joi.string().required(),
-          thumbnail: Joi.string().required() ,
-          tempImg: Joi.string().allow(null).empty(''),
+          thumbnail: Joi.allow(null).empty(''),
           category: Joi.string().required(),
           content: Joi.string().required(),
+          count: Joi.string().required(),
+          metaTitle: Joi.allow(null).empty(''),
+          metaDescription: Joi.allow(null).empty(''),
+          metaKeywords: Joi.allow(null).empty(''),
         });
-        const { error } = tipSchema.validate(req.body);
+        const { error } = blogSchema.validate(req.body);
         
         // 2. if error in validation -> return error via middleware
         if (error) {
@@ -105,58 +26,138 @@ const tipsController = {
         }
         
         // 3. if email or username is already registered -> return an error
-        const {id,title,slug,thumbnail,tempImg,category,content } = req.body;
+        const {title,slug,category,content,count,metaTitle,metaDescription,metaKeywords } = req.body;
         
-        const inUse = await Tip.exists({ _id:id });        
-        if (!inUse) {
+        const slugInUse = await Tip.exists({ slug });        
+        if (slugInUse) {
           const error = {
-            status: 409, message:'Appliance Tip Not Found!'
+            status: 409, message:'Tip Already Exist!'
           }
           return next(error)
         }
+        
+        let img='';
+        if(req.files?.thumbnail){
+          const {response,updateImg} = await AWSService.uploadFile({name:req.files.thumbnail.name,data:req.files.thumbnail.data},'blog/')
+         if(response.$metadata.httpStatusCode === 200){
+          img = updateImg;
+         }else{
+           const error = {status: 500,message: "AWS S3 Internal Server Error!"}
+           return next(error)
+         }
+       }
 
-        if(tempImg === ''){
-          try {
-              const updatedBlog = await Tip.findByIdAndUpdate(
-                id,
-                {title,slug,thumbnail,category,content},
-                { new: true }
-              );
-      
-              return res.status(200).json({status:200,msg:'Appliance Tip Updated Successfully!'});
-          
-              } catch (error) {
-                const err = {status:500,msg:"Internal Server Error!"}
-                return next(err);
-              }
-        }else{
-          const {resp} = await AWSService.duplicateFile(thumbnail)
-          try{
-           if(resp.$metadata.httpStatusCode === 200){
-             const {response,updateImg} = await AWSService.uploadFile({name:req.files.thumbnail.name,data:req.files.thumbnail.data},'appliance-tips/')
-             try{
-               if(response.$metadata.httpStatusCode === 200){
-                try{
-                  const TipToCreate = new Tip({title,slug,thumbnail:updateImg,category,content});
-                  const tip = await TipToCreate.save();
-                  return res.status(200).json({status: 200, msg:'Appliance Tip Updated Successuly!'});
-                 }catch(err){
-                   const error = {status:500,msg:"Internal Server Error!"}
-                   return next(error)
-                 }
-               }
-             }catch(error){
-               const err = {status:500,message:"Cloud Internal Server Server!"} 
-               return next(err)
-             }
-
-           }
-          }catch(error){
-           const err = {status:500,message:"Cloud Internal Server Server!"} 
-           return next(err)
-          }
-        }
+       try{
+        const BlogToCreate = new Tip({title,slug,count,thumbnail:img,category,content,metaTitle,metaDescription,metaKeywords});
+        await BlogToCreate.save();
+        return res.status(200).json({status: 200, msg:'Tip Created Successuly!'});
+       }catch(err){
+         const error = {status:500,massage:"Internal Server Error!"}
+         return next(error)
+       }
   },
+
+  async duplicateTip(req, res, next) {
+    const blogSchema = Joi.object({
+        slug: Joi.string().required(),
+      });
+      const { error } = blogSchema.validate(req.body);
+      
+      // 2. if error in validation -> return error via middleware
+      if (error) {
+        return next(error)
+      }
+      // 
+      // 3. if email or username is already registered -> return an error
+      const {slug} = req.body;
+      
+      let img ='';
+      const blog = await Tip.findOne({ slug });  
+      if (blog) {
+       if(blog.thumbnail){
+        const {resp,updateImg} = await AWSService.duplicateFile(blog.thumbnail,'tip/')
+        if(resp.$metadata.httpStatusCode === 200){
+         img = updateImg;
+        }else{
+          return res.status(500).json({status: 500, message:'AWS Cloud Internal Server Error!'});
+        }
+       }
+       const title = blog.title + '(duplicate)';
+       const slug = title.toLocaleLowerCase().replace(/\s/g,'-');
+       try{
+         const BlogToCreate = new Tip({title:title,slug:slug,thumbnail:img,count:blog.count,category:blog.category,content:blog.content,metaTitle:blog.metaTitle,metaDescription:blog.metaDescription,metaKeywords:blog.metaKeywords});
+         await BlogToCreate.save();
+         return res.status(200).json({status: 200, msg:'Blog Duplicate Created!'});
+       }catch(err){
+         const error = {status:500,message:"Internal Server Error!"}
+         return next(error)
+       }
+      }else{
+        return res.status(404).json({status: 404, message:'Blog Does Not Exist!'});
+      }  
+},
+
+async updateTip(req, res, next) {
+  const blogSchema = Joi.object({
+      id: Joi.string().required(),
+      title: Joi.string().required(),
+      slug: Joi.string().required(),
+      thumbnail: Joi.string().allow(null).empty(''),
+      tempImg: Joi.string().allow(null).empty(''),
+      category: Joi.string().required(),
+      isImg: Joi.boolean().required(),
+      content: Joi.string().required(),
+      count: Joi.string().required(),
+      metaTitle: Joi.string().allow(null).empty(''),
+      metaDescription: Joi.string().allow(null).empty(''),
+      metaKeywords: Joi.string().allow(null).empty(''),
+    });
+    const { error } = blogSchema.validate(req.body);
+    
+    // 2. if error in validation -> return error via middleware
+    if (error) {
+      return next(error)
+    }
+    
+    // 3. if email or username is already registered -> return an error
+    const {id,title,slug,thumbnail,count,tempImg,isImg,category,content,metaTitle,metaDescription,metaKeywords } = req.body;
+    const inUse = await Tip.exists({ _id:id });        
+    if (!inUse) {
+      const error = {
+        status: 409, message:'Tip Not Found!'
+      }
+      return next(error)
+    }
+
+
+    let img = thumbnail;
+    if(isImg === 'true'){
+     if(tempImg){
+      const {resp} = await AWSService.deleteFile(tempImg)
+      if(resp.$metadata.httpStatusCode !== 200){
+        const err = {status:500,message:"Cloud Internal Server Server!"};return next(err)
+      }
+     }else{
+      const {response,updateImg} = await AWSService.uploadFile({name:req.files.thumbnail.name,data:req.files.thumbnail.data},'blog/')
+      if(response.$metadata.httpStatusCode === 200){
+       img = updateImg;
+      }else{
+        const err = {status:500,message:"Cloud Internal Server Server!"};return next(err)
+      }
+     }
+    }
+
+    // try {
+    await Tip.findByIdAndUpdate(
+      {_id:id},
+      {title,slug,thumbnail:img,count,category,content,metaTitle,metaDescription,metaKeywords:JSON.parse(metaKeywords)},);
+     return res.status(200).json({status:200,msg:'Tip Updated Successfully!'});
+    // } catch (error) {
+    //   const err = {status:500,message:"Internal Server Error!"}
+    //   return next(err);
+    // }
+},
+
     async getRecentTips(req, res, next) {
 
         try{
@@ -194,7 +195,7 @@ const tipsController = {
       const {slug} = req.body;
 
       try{
-        const tip = await Tip.find({slug:slug});        
+        const tip = await Tip.findOne({slug:slug});        
         return res.status(200).json({status: 200, tip});
       }catch(error){
         return next(error)
@@ -254,25 +255,27 @@ async deleteTip(req, res, next) {
     }
 
     const {id} = req.body;
-    
+
+    let tip;
     try{
-     const tip = await Tip.findByIdAndDelete(id);
-     
+     tip = await Tip.findOne({_id:id})
      if(!tip){
-       return res.status(500).json({status: 500, message:'Appliance Tip Not Found!'});
+      return res.status(404).json({status:404,message:'Appliance Tip Not Found!'})
      }
-       try{
-         const {resp} = await AWSService.deleteFile(tip.thumbnail)
-         if(resp.$metadata.httpStatusCode === 204){
-           return res.status(200).json({status: 200, msg:'Appliance Tip Deleted Successfully!'});    
-         }
-        }catch(error){
-         const err = {status:500,message:"Cloud Internal Server Server!"} 
-         return next(err)
-        }
-      }catch(error){
-       return next(error)
+    }catch(error){return res.status(500).json({status:500,message:'Internal Server Error!'})}
+    
+    if(tip.thumbnail){
+      const {resp} = await AWSService.deleteFile(tip.thumbnail)
+      if(resp.$metadata.httpStatusCode !== 204){
+        const err = {status:500,message:"Cloud Internal Server Server!"} 
+        return next(err)
       }
+    }
+
+    try{
+      await Tip.findByIdAndDelete(id);
+      return res.status(200).json({status: 200, msg:'Appliance Tip Deleted Successfully!'});    
+    }catch(error){return res.status(500).json({status:500,message:'Internal Server Error!'})}
 },
 async getTipBySearch(req, res, next) {
   const tipSchema = Joi.object({
@@ -304,8 +307,35 @@ async getTipBySearch(req, res, next) {
       return next(error)
     }
 
-}
+},
 
+ async getFourTips(req, res, next) {
+     try{
+       const tips = await Tip.find({}).limit(6); 
+       const totalCount = await Tip.countDocuments();
+       return res.status(200).json({status: 200, tips:tips,totalCount:totalCount});
+     }catch(error){
+       return next(error)
+     }
+ },
+
+
+ async GetTipBySlugWithCategories(req,res,next){
+
+  let blog = {}
+  let categories = []
+  try{
+    categories = await Category.find({}).select('title').select('slug');
+  }catch(error){
+    return next(error)
+  }
+  try{
+    blog = await Tip.findOne({slug:req.body.slug});
+  }catch(error){
+    return next(error)
+  }
+  return res.status(200).json({status:200,categories:categories,blog:blog});
+},
 
 }
